+++ doajtest/fixtures/background.py
@@ -1,6 +1,10 @@
 from copy import deepcopy
-from portality.background import BackgroundTask
+
 from portality import models
+from portality.background import BackgroundTask
+from portality.lib import dates
+from portality.models import BackgroundJob
+
 
 class BackgroundFixtureFactory(object):
 
@@ -31,18 +35,40 @@ class MockBackgroundTask(BackgroundTask):
             raise Exception("Mock Cleanup Failed")
 
 
-
 BACKGROUND_JOB = {
-    "id" : "123456789",
-    "created_date" : "2001-01-01T00:00:00Z",
-    "last_updated" : "2001-01-02T00:00:00Z",
-    "status" : "queued",
-    "user" : "testuser",
-    "queue_id" : "abcdef",
-    "action" : "status_change",
-    "params" : {},
-    "reference" : {},
-    "audit" : [
-        {"message" : "created job", "timestamp" : "2001-01-01T00:00:00Z"}
+    "id": "123456789",
+    "created_date": "2001-01-01T00:00:00Z",
+    "last_updated": "2001-01-02T00:00:00Z",
+    "status": "queued",
+    "user": "testuser",
+    "queue_id": "abcdef",
+    "action": "status_change",
+    "params": {},
+    "reference": {},
+    "audit": [
+        {"message": "created job", "timestamp": "2001-01-01T00:00:00Z"}
     ]
 }
+
+
+def save_mock_bgjob(action=None, status=None, created_before_sec=0, is_save=True,
+                    queue_type=None):
+    bgjob = BackgroundJob()
+
+    if action:
+        from portality.tasks.journal_csv import JournalCSVBackgroundTask
+        bgjob.action = JournalCSVBackgroundTask.__action__
+
+    if status:
+        bgjob._set_with_struct("status", status)
+
+    if created_before_sec != 0:
+        bgjob.set_created(dates.format(dates.before_now(created_before_sec)))
+
+    if queue_type:
+        bgjob.queue_type = queue_type
+
+    if is_save:
+        bgjob.save()
+
+    return bgjob
